# docs/swagger.yaml
basePath: /api/v1
definitions:
  barber-booking-system_internal_middleware.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: true
        type: object
      error:
        type: string
      message:
        type: string
    type: object
...
  barber-booking-system_internal_services.UpdateReviewRequest:
    properties:
      cleanliness_rating:
        maximum: 5
        minimum: 1
        type: integer
      comment:
        type: string
      cons:
        type: string
      duration_accurate:
        type: boolean
      images:
        items:
          type: string
        type: array
      overall_rating:
        maximum: 5
        minimum: 1
        type: integer
      professionalism_rating:
        maximum: 5
        minimum: 1
        type: integer
      pros:
        type: string
      punctuality_rating:
        maximum: 5
        minimum: 1
        type: integer
      service_as_expected:
        type: boolean
      service_quality_rating:
        maximum: 5
        minimum: 1
        type: integer
      title:
        type: string
      value_for_money_rating:
        maximum: 5
        minimum: 1
        type: integer
      would_book_again:
        type: boolean
      would_recommend:
        type: boolean
    type: object
...
  internal_handlers.SuccessResponse:
    description: Standard success response wrapper
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      meta: {}
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@barbershop.com
    name: API Support
  description: |-
    A production-ready RESTful API for barbershop booking system.
    Features include: User authentication, Barber management, Service catalog, Booking system, Review & Rating system, and Notification system.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Barbershop Booking API
  version: 1.0.0
paths:
...
  /api/v1/barbers/{id}/reviews:
    get:
      consumes:
      - application/json
      description: Get all published reviews for a specific barber
      parameters:
      - description: Barber ID
        in: path
        name: id
        required: true
        type: integer
      - description: Filter by minimum rating
        in: query
        name: min_rating
        type: integer
      - description: Filter by maximum rating
        in: query
        name: max_rating
        type: integer
      - description: Filter reviews with comments
        in: query
        name: has_comment
        type: boolean
      - description: Filter reviews with images
        in: query
        name: has_images
        type: boolean
      - default: created_at
        description: Sort by field
        in: query
        name: sort_by
        type: string
      - default: DESC
        description: Sort order (ASC/DESC)
        in: query
        name: order
        type: string
      - default: 50
        description: Limit results
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      summary: Get barber's reviews
      tags:
      - reviews
  /api/v1/barbers/{id}/reviews/stats:
    get:
      consumes:
      - application/json
      description: Get review statistics (ratings distribution, average) for a barber
      parameters:
      - description: Barber ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      summary: Get barber's review statistics
      tags:
      - reviews
...
  /api/v1/reviews:
    post:
      consumes:
      - application/json
      description: Create a review for a completed booking
      parameters:
      - description: Review data
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/barber-booking-system_internal_services.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "409":
          description: Review already exists
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new review
      tags:
      - reviews
  /api/v1/reviews/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a review (soft delete by author or admin)
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a review
      tags:
      - reviews
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific review
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      summary: Get review by ID
      tags:
      - reviews
    put:
      consumes:
      - application/json
      description: Update a review (only by author and before moderation)
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated review data
        in: body
        name: review
        required: true
        schema:
          $ref: '#/definitions/barber-booking-system_internal_services.UpdateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a review
      tags:
      - reviews
  /api/v1/reviews/{id}/moderate:
    patch:
      consumes:
      - application/json
      description: Update the moderation status of a review (admin only)
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      - description: Moderation action
        in: body
        name: moderation
        required: true
        schema:
          $ref: '#/definitions/barber-booking-system_internal_services.ModerateReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Moderate a review
      tags:
      - reviews
  /api/v1/reviews/{id}/response:
    post:
      consumes:
      - application/json
      description: Allow a barber to respond to a review on their profile
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      - description: Response content
        in: body
        name: response
        required: true
        schema:
          $ref: '#/definitions/barber-booking-system_internal_services.BarberResponseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Add barber response to a review
      tags:
      - reviews
  /api/v1/reviews/{id}/vote:
    post:
      consumes:
      - application/json
      description: Vote whether a review was helpful or not
      parameters:
      - description: Review ID
        in: path
        name: id
        required: true
        type: integer
      - description: Vote data
        in: body
        name: vote
        required: true
        schema:
          $ref: '#/definitions/barber-booking-system_internal_services.VoteReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      summary: Vote on review helpfulness
      tags:
      - reviews
  /api/v1/reviews/booking/{booking_id}:
    get:
      consumes:
      - application/json
      description: Get the review for a specific booking
      parameters:
      - description: Booking ID
        in: path
        name: booking_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/barber-booking-system_internal_middleware.ErrorResponse'
      summary: Get review by booking ID
      tags:
      - reviews
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...